requirements.txt
pandas>=2.2.3
torch>=2.6.0
transformers>=4.50.3
scikit-learn>=1.6.1
python-pptx>=1.0.2 